+++ service/src/main/java/com/gmail/nogovitsyndmitriy/service/impl/UserServiceImpl.java
@@ -8,6 +8,7 @@ import com.gmail.nogovitsyndmitriy.dao.entities.User;
 import com.gmail.nogovitsyndmitriy.service.UserService;
 import com.gmail.nogovitsyndmitriy.service.converter.impl.dto.UserDtoConverter;
 import com.gmail.nogovitsyndmitriy.service.converter.impl.entity.UserConverter;
+import com.gmail.nogovitsyndmitriy.service.exceptions.ServiceException;
 import com.gmail.nogovitsyndmitriy.service.model.PasswordDto;
 import com.gmail.nogovitsyndmitriy.service.model.UserDto;
 import org.apache.logging.log4j.LogManager;
@@ -18,6 +19,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -30,31 +32,30 @@ public class UserServiceImpl implements UserService {
     private final RoleDao roleDao;
     private final BCryptPasswordEncoder encoder;
 
-
     @Autowired
     public UserServiceImpl(@Qualifier("userDtoConverter") UserDtoConverter userDtoConverter,
                            @Qualifier("userConverter") UserConverter userConverter,
                            UserDao userDao,
                            RoleDao roleDao,
-                           BCryptPasswordEncoder encoder) {
+                           BCryptPasswordEncoder encoder
+    ) {
         this.userDtoConverter = userDtoConverter;
         this.userConverter = userConverter;
         this.userDao = userDao;
         this.roleDao = roleDao;
         this.encoder = encoder;
-
     }
 
     @Override
     @Transactional(readOnly = true)
     public UserDto get(Long id) {
-        UserDto userDto = new UserDto();
-        try {
-            User user = userDao.get(id);
+        UserDto userDto;
+        User user = userDao.get(id);
+        if (user != null) {
             userDto = userDtoConverter.toDTO(user);
             log.info("Get user successful!");
-        } catch (Exception e) {
-            log.error("Get user failed!", e);
+        } else {
+            throw new EntityNotFoundException("Entity with id:" + id + " not found!");
         }
         return userDto;
     }
@@ -75,6 +76,7 @@ public class UserServiceImpl implements UserService {
             log.info("Saving user successful!");
         } catch (Exception e) {
             log.error("Saving user failed!", e);
+            throw new ServiceException("666", "Service Exception!");
         }
         return userDto;
     }
@@ -89,6 +91,7 @@ public class UserServiceImpl implements UserService {
             log.info("Update user successful!");
         } catch (Exception e) {
             log.error("Update user failed!", e);
+            throw new ServiceException("666", "Service Exception!");
         }
         return userDto;
     }
@@ -102,30 +105,32 @@ public class UserServiceImpl implements UserService {
             log.info("Delete user successful!");
         } catch (Exception e) {
             log.error("Delete user failed!", e);
+            throw new ServiceException("666", "Service Exception!");
         }
     }
 
     @Override
     @Transactional
     public void deleteById(Long id) {
-        try {
-            User user = userDao.get(id);
+        User user = userDao.get(id);
+        if (user != null) {
             userDao.delete(user);
             log.info("Delete user by Id successful!");
-        } catch (Exception e) {
-            log.error("Delete user by Id failed!", e);
+        } else {
+            throw new EntityNotFoundException("Item NOT Found!");
         }
     }
 
     @Override
     @Transactional
     public List<UserDto> getAll() {
-        List<UserDto> users = new ArrayList<>();
+        List<UserDto> users;
         try {
             users = userDtoConverter.toDtoList(userDao.getAll());
             log.info("Get all users successful!");
         } catch (Exception e) {
             log.info("Get all users failed!");
+            throw new ServiceException("666", "Service Exception!");
         }
         return users;
     }
@@ -134,13 +139,14 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional(readOnly = true)
     public UserDto findByEmail(String email) {
-        UserDto userDto = new UserDto();
+        UserDto userDto;
         try {
             User user = userDao.findByEmail(email);
             userDto = userDtoConverter.toDTO(user);
             log.info("Find user by email successful!");
         } catch (Exception e) {
             log.error("Find user by email failed!", e);
+            throw new ServiceException("666", "Service Exception!");
         }
         return userDto;
     }
@@ -148,12 +154,13 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional(readOnly = true)
     public Long quantityOfUsers() {
-        Long quantity = 0L;
+        Long quantity;
         try {
             quantity = userDao.quantityOfUsers();
             log.info("Quantity find successful!");
         } catch (Exception e) {
             log.error("Failed to count quantity!", e);
+            throw new ServiceException("666", "Service Exception!");
         }
         return quantity;
     }
@@ -171,6 +178,7 @@ public class UserServiceImpl implements UserService {
             log.info("User pangination successful!");
         } catch (Exception e) {
             log.error("Failed to get users pangination");
+            throw new ServiceException("666", "Service Exception!");
         }
         return usersDto;
     }
@@ -183,13 +191,13 @@ public class UserServiceImpl implements UserService {
             password.setNewPassword(encoder.encode(password.getNewPassword()));
             password.setConfirmPassword(encoder.encode(password.getConfirmPassword()));
             userDto.setPassword(password.getNewPassword());
-
             User user = userConverter.toEntity(userDto);
             userDao.update(user);
             userDto = userDtoConverter.toDTO(user);
             return userDto;
+        } else {
+            throw new ServiceException("666", "Wrong password!");
         }
-        return null;
     }
 
     @Override
